+++ src/util/pixuntile.c
@@ -1,7 +1,7 @@
 /*                     P I X U N T I L E . C
  * BRL-CAD
  *
- * Copyright (c) 1986-2010 United States Government as represented by
+ * Copyright (c) 1986-2011 United States Government as represented by
  * the U.S. Army Research Laboratory.
  *
  * This program is free software; you can redistribute it and/or
@@ -124,7 +124,7 @@ main(int argc, char **argv)
     }
 
     if (in_width < 1) {
-	fprintf(stderr, "pixuntile: width of %lu out of range\n", in_width);
+	fprintf(stderr, "pixuntile: width of %lu out of range\n", (long unsigned)in_width);
 	bu_exit (12, NULL);
     }
